Add sell header to Bag
